Доработка модуля обновления цен/остатков, добавлены монобренды (таблица, возможность обновления)

+++ EtkBlazorApp/Model/PrikatManufacturerDiscount.cs
@@ -7,8 +7,8 @@ namespace EtkBlazorApp.ViewModel
     public class PrikatManufacturerDiscountViewModel
     {       
         public int Manufacturer_id { get; set; }
-        public int Discount1 { get; set; }
-        public int Discount2 { get; set; }
+        public decimal Discount1 { get; set; }
+        public decimal Discount2 { get; set; }
         public bool IsChecked { get; set; }
         public string Manufacturer { get; set; }
         public string CurrencyCode { get; set; }
